requirements.txt
Flask
Gunicorn
supabase
python-dotenv
flask-cors